envelop_0.5.2/envelop/requirements.txt
Django==2.1.4
django-registration
python-dateutil